rename loaderInterface.py to loader_interface.py
